Shader rewrite
I absolutely hate all of this.
All of it.
Was this worth it? Absolutely not.
Why did i do it? Because i like "progress" or whatever and don't like to use old stuff (In other words i hate myself)

Greedy meshing is still not ported.
Chunk loading/unloading still not ported.
Block highlight still not ported
HUD might be a bit complicated to port..?
Profiling and FPS metrics still not ported.

This rewrite is a 82.5% perf decrease, Nice.